requirements.txt
google-genai>=0.5.0
sqlalchemy>=2.0
pandas>=2.0
streamlit>=1.32
python-dotenv>=1.0
plotly>=5.20
faker>=24.0
openpyxl>=3.1
